Create tool-the-cursor1-v1
To make Cursor No. 1 more practical, I adapted it: 
this version ("v1") starts out as the original, with a trail covering the information of the respective page or post or dataset.

But for the sake of usability, in this version, the trail-effect STOPS AFTER 15 SECONDS – and morphs into a semi-translucent light blue dot.

That way, users can proceed with their research, while being reminded of the original effect at the same time.